unstructured/requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
pypdf==4.2.0
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.2.1
pandas==2.2.2
Pillow==10.3.0
pytesseract==0.3.10
unstructured==0.14.10
chromadb==0.4.24
sentence-transformers==2.5.1
python-dotenv==1.0.1
fastapi==0.110.2
uvicorn==0.29.0
